pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "google-api-python-client>=2.183.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "insightface>=0.7.3",
    "numpy>=2.3.1",
    "onnxruntime>=1.22.0",
    "opencv-python>=4.11.0.86",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
]
